bot/pyproject.toml
[project]
name = "cryptomacro-bot"
version = "0.1.0"
description = "Discord bot for alert delivery and slash commands"
requires-python = ">=3.11"

dependencies = [
    "discord.py>=2.3.0",
    "nats-py>=2.9.0",
    "psycopg[binary]>=3.1.0",
    "psycopg-pool>=3.1.0",
    "redis>=5.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "structlog>=24.0.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["src"]
